requirements.txt
nicegui>=2.0
pillow>=10.0
pywebview>=5.0; platform_system == "Windows"
